+++ EdkCompatibilityPkg/Foundation/Library/EdkIIGlueLib/Library/BaseLib/MultS64x64.c
@@ -45,5 +45,5 @@ MultS64x64 (
   IN      INT64                     Multiplier
   )
 {
-  return (INT64)MultU64x64 (Multiplicand, Multiplier);
+  return (INT64)MultU64x64 ((UINT64) Multiplicand, (UINT64) Multiplier);
 }
