+++ EdkCompatibilityPkg/Foundation/Library/EdkIIGlueLib/Library/DxeReportStatusCodeLib/ReportStatusCodeLib.c
@@ -512,7 +512,7 @@ GlueReportStatusCodeEx (
   //
   // Fill in the extended data header
   //
-  StatusCodeData->HeaderSize = sizeof (EFI_STATUS_CODE_DATA);
+  StatusCodeData->HeaderSize = (UINT16) sizeof (EFI_STATUS_CODE_DATA);
   StatusCodeData->Size = (UINT16)ExtendedDataSize;
   if (ExtendedDataGuid == NULL) {
     ExtendedDataGuid = &gEfiStatusCodeSpecificDataGuid;
