+++ EdkCompatibilityPkg/Foundation/Library/EdkIIGlueLib/Library/PeiReportStatusCodeLib/ReportStatusCodeLib.c
@@ -447,7 +447,7 @@ GlueReportStatusCodeEx (
     return EFI_OUT_OF_RESOURCES;
   }
   StatusCodeData = (EFI_STATUS_CODE_DATA  *)Buffer;
-  StatusCodeData->HeaderSize = sizeof (EFI_STATUS_CODE_DATA);
+  StatusCodeData->HeaderSize = (UINT16) sizeof (EFI_STATUS_CODE_DATA);
   StatusCodeData->Size = (UINT16)ExtendedDataSize;
   if (ExtendedDataGuid == NULL) {
     ExtendedDataGuid = &gEfiStatusCodeSpecificDataGuid;
